requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
billiard==3.6.4.0
black==23.1.0
boto3==1.24.59
botocore==1.27.59
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.1.4
dulwich==0.21.2
dvc==2.43.4
dvc-data==0.37.8
dvc-http==2.30.2
dvc-objects==0.19.3
dvc-render==0.1.0
dvc-s3==2.21.0
dvc-studio-client==0.4.0
dvc-task==0.1.11
dvclive==2.0.2
filelock==3.9.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
funcy==1.18
gitdb==4.0.10
GitPython==3.1.30
grandalf==0.8
hydra-core==1.3.1
idna==3.4
importlib-resources==5.12.0
iterative-telemetry==0.0.7
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
markdown-it-py==2.1.0
matplotlib==3.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nanotime==0.5.2
networkx==3.0
numpy==1.24.2
omegaconf==2.3.0
orjson==3.8.5
packaging==23.0
pandas==1.5.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rich==13.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2023.1.0
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
scmrepo==0.1.7
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sqltrie==0.0.28
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
wrapt==1.14.1
yarl==1.8.2
zc.lockfile==2.0
zipp==3.14.0
